+++ src/ssi.h
@@ -1,6 +1,6 @@
 /* Copyright (C) 1995, 1996 by Sven Berkvens (sven@stack.nl) */
 
 #ifdef		WANT_SSI
-int	sendwithdirectives	(int, size_t *);
+int	sendwithdirectives	(int, off_t *);
 int	counter_versioncheck	(void);
 #endif		/* WANT_SSI */
